+++ toteatds/src/commonMain/kotlin/com/toteat/toteatds/components/tags/StatusTagVariant.kt
@@ -1,6 +1,8 @@
 package com.toteat.toteatds.components.tags
 
 import androidx.compose.ui.graphics.Color
+import com.toteat.toteatds.theme.BlueLight
+import com.toteat.toteatds.theme.BlueNormal
 import com.toteat.toteatds.theme.NeutralGray
 import com.toteat.toteatds.theme.NeutralGray200
 import com.toteat.toteatds.theme.NeutralGray400
@@ -10,6 +12,7 @@ import designsystemmobile.toteatds.generated.resources.Res
 import designsystemmobile.toteatds.generated.resources.status_tag_cancelled
 import designsystemmobile.toteatds.generated.resources.status_tag_confirmed
 import designsystemmobile.toteatds.generated.resources.status_tag_ended
+import designsystemmobile.toteatds.generated.resources.status_tag_in_promotion
 import designsystemmobile.toteatds.generated.resources.status_tag_pending
 import org.jetbrains.compose.resources.StringResource
 
@@ -37,5 +40,10 @@ enum class StatusTagVariant(
         backgroundColor = SecondaryNormal,
         textColor = NeutralGray,
         defaultTextRes = Res.string.status_tag_cancelled
+    ),
+    Promotion(
+        backgroundColor = BlueLight,
+        textColor = BlueNormal,
+        defaultTextRes = Res.string.status_tag_in_promotion
     )
 }